split linear_expression.py
