+++ veerer/polyhedron/__init__.py
@@ -1,0 +1,1 @@
+from .linear_expression import LinearExpressions, ConstraintSystem
