+++ veerer/polyhedron/linear_algebra.py
